+++ fabric-common-tags-v1/build.gradle
@@ -1,4 +1,4 @@
-archivesBaseName = "fabric-common-tags-v1"
+archivesBaseName = "fabric-convention-tags-v1"
 version = getSubprojectVersion(project)
 
 moduleDependencies(project, [
