+++ fabric-common-tags-v1/src/datagen/java/net/fabricmc/fabric/impl/tag/convention/datagen/generators/BiomeTagGenerator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.tag.datagen.generators;
+package net.fabricmc.fabric.impl.tag.convention.datagen.generators;
 
 import net.minecraft.tag.BiomeTags;
 import net.minecraft.util.registry.Registry;
@@ -23,7 +23,7 @@ import net.minecraft.world.biome.BiomeKeys;
 
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
-import net.fabricmc.fabric.api.tag.v1.CommonBiomeTags;
+import net.fabricmc.fabric.api.tag.convention.v1.CommonBiomeTags;
 
 public class BiomeTagGenerator extends FabricTagProvider.DynamicRegistryTagProvider<Biome> {
 	public BiomeTagGenerator(FabricDataGenerator dataGenerator) {
