+++ fabric-common-tags-v1/src/datagen/java/net/fabricmc/fabric/impl/tag/convention/datagen/generators/BlockTagGenerator.java
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.tag.datagen.generators;
+package net.fabricmc.fabric.impl.tag.convention.datagen.generators;
 
 import net.minecraft.block.Blocks;
 import net.minecraft.tag.BlockTags;
 
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
-import net.fabricmc.fabric.api.tag.v1.CommonBlockTags;
+import net.fabricmc.fabric.api.tag.convention.v1.CommonBlockTags;
 
 public class BlockTagGenerator extends FabricTagProvider.BlockTagProvider {
 	public BlockTagGenerator(FabricDataGenerator dataGenerator) {
