+++ fabric-common-tags-v1/src/datagen/java/net/fabricmc/fabric/impl/tag/convention/datagen/generators/FluidTagGenerator.java
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.tag.datagen.generators;
+package net.fabricmc.fabric.impl.tag.convention.datagen.generators;
 
 import net.minecraft.tag.FluidTags;
 
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
-import net.fabricmc.fabric.api.tag.v1.CommonFluidTags;
+import net.fabricmc.fabric.api.tag.convention.v1.CommonFluidTags;
 
 public class FluidTagGenerator extends FabricTagProvider.FluidTagProvider {
 	public FluidTagGenerator(FabricDataGenerator dataGenerator) {
