+++ fabric-common-tags-v1/src/datagen/java/net/fabricmc/fabric/impl/tag/convention/datagen/generators/ItemTagGenerator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.tag.datagen.generators;
+package net.fabricmc.fabric.impl.tag.convention.datagen.generators;
 
 import net.minecraft.item.Items;
 import net.minecraft.tag.ItemTags;
@@ -22,7 +22,7 @@ import net.minecraft.util.Identifier;
 
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
-import net.fabricmc.fabric.api.tag.v1.CommonItemTags;
+import net.fabricmc.fabric.api.tag.convention.v1.CommonItemTags;
 
 public class ItemTagGenerator extends FabricTagProvider.ItemTagProvider {
 	/**
