+++ fabric-common-tags-v1/src/datagen/resources/fabric.mod.json
@@ -1,11 +1,11 @@
 {
   "schemaVersion": 1,
-  "id": "fabric-common-tags-v1-datagen",
-  "name": "Fabric Common tags datagen",
+  "id": "fabric-convention-tags-v1-datagen",
+  "name": "Fabric Convention Tags Datagen",
   "version": "1.0.0",
   "environment": "*",
   "license": "Apache-2.0",
-  "icon": "assets/fabric-common-tags-v1/icon.png",
+  "icon": "assets/fabric-convention-tags-v1/icon.png",
   "contact": {
     "homepage": "https://fabricmc.net",
     "irc": "irc://irc.esper.net:6667/fabric",
@@ -25,7 +25,7 @@
   },
   "entrypoints": {
     "fabric-datagen": [
-      "net.fabricmc.fabric.impl.tag.common.datagen.DatagenEntrypoint"
+      "net.fabricmc.fabric.impl.tag.convention.datagen.DatagenEntrypoint"
     ]
   }
 }
