+++ fabric-common-tags-v1/src/main/java/net/fabricmc/fabric/api/tag/convention/v1/CommonBlockTags.java
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.api.tag.v1;
+package net.fabricmc.fabric.api.tag.convention.v1;
 
 import net.minecraft.block.AbstractBlock;
 import net.minecraft.block.Block;
 import net.minecraft.tag.TagKey;
 
-import net.fabricmc.fabric.impl.tag.common.TagRegistration;
+import net.fabricmc.fabric.impl.tag.convention.TagRegistration;
 
 /**
  * See {@link net.minecraft.tag.BlockTags} for vanilla tags.
