+++ fabric-common-tags-v1/src/main/java/net/fabricmc/fabric/api/tag/convention/v1/CommonEnchantmentTags.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.api.tag.v1;
+package net.fabricmc.fabric.api.tag.convention.v1;
 
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.tag.TagKey;
 
-import net.fabricmc.fabric.impl.tag.common.TagRegistration;
+import net.fabricmc.fabric.impl.tag.convention.TagRegistration;
 
 public class CommonEnchantmentTags {
 	/**
