+++ fabric-common-tags-v1/src/main/resources/assets/fabric-convention-tags-v1/icon.png
