+++ gradle.properties
@@ -51,4 +51,4 @@ fabric-screen-handler-api-v1-version=1.1.14
 fabric-textures-v0-version=1.0.12
 fabric-transfer-api-v1-version=1.5.14
 fabric-transitive-access-wideners-v1-version=1.0.0
-fabric-common-tags-v1-version=1.0.0
+fabric-convention-tags-v1-version=1.0.0
