+++ settings.gradle
@@ -54,5 +54,5 @@ include 'fabric-screen-api-v1'
 include 'fabric-screen-handler-api-v1'
 include 'fabric-textures-v0'
 include 'fabric-transfer-api-v1'
-include 'fabric-common-tags-v1'
+include 'fabric-convention-tags-v1'
 include 'fabric-transitive-access-wideners-v1'
